ISSUE-538 Move transaction package from common to storage-core
