+++ storage/core/src/main/java/com/hortonworks/registries/storage/impl/jdbc/provider/sql/factory/QueryExecutor.java
@@ -16,7 +16,7 @@
 
 package com.hortonworks.registries.storage.impl.jdbc.provider.sql.factory;
 
-import com.hortonworks.registries.common.transaction.TransactionIsolation;
+import com.hortonworks.registries.storage.transaction.TransactionIsolation;
 import com.hortonworks.registries.storage.OrderByField;
 import com.hortonworks.registries.storage.Storable;
 import com.hortonworks.registries.storage.StorableFactory;
