+++ storage/core/src/main/java/com/hortonworks/registries/storage/transaction/TransactionIsolation.java
@@ -1,4 +1,4 @@
-package com.hortonworks.registries.common.transaction;
+package com.hortonworks.registries.storage.transaction;
 
 import java.sql.Connection;
