+++ storage/core/src/test/java/com/hortonworks/registries/storage/filestorage/DbFileStorageTest.java
@@ -15,7 +15,7 @@
  **/
 package com.hortonworks.registries.storage.filestorage;
 
-import com.hortonworks.registries.common.transaction.TransactionIsolation;
+import com.hortonworks.registries.storage.transaction.TransactionIsolation;
 import com.hortonworks.registries.storage.StorageManager;
 import com.hortonworks.registries.storage.TransactionManager;
 import com.hortonworks.registries.storage.exception.StorageException;
